refactor: Split add vector service
Dynamically load some of OL format classes
